API fixes

+++ aspnet-core/src/LoanManagement.Application/DatabaseServices/Interfaces/IMortageService.cs
@@ -3,7 +3,7 @@ using LoanManagement.ViewModels;
 
 namespace LoanManagement.DatabaseServices.Interfaces
 {
-    public interface IMortageService : IAsyncCrudAppService<MortgageTypeDto, int, PagedLoanApplicationResultRequestDto, MortgageTypeDto, MortgageTypeDto>
+    public interface IMortageService : IAsyncCrudAppService<MortgageTypeDto, long, PagedLoanApplicationResultRequestDto, MortgageTypeDto, MortgageTypeDto>
     {
     }
 }
